pyproject.toml
[tool.poetry]
name = "lfr"
version = "0.1.3"
description = ""
authors = ["Radhakrishna Sanka <rkrishnasanka@gmail.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.8"
networkx = "^2.5"
argparse = "^1.4.0"
antlr4-python3-runtime = "^4.8"
pygraphviz = "^1.6"
pymint = "^0.1.3"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
black = "^20.8b1"
flake8 = "^3.8.3"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
lfr-compile = "lfr.cmdline:main"
